% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACY.R
\name{acy_sim}
\alias{acy_sim}
\title{Simulate ACY processes}
\usage{
acy_sim(story = NULL, ..., samp_n = 5, numerical = "Y")
}
\arguments{
\item{story}{Optional list containing formulas and/or label names}

\item{...}{more formulas describing the process and/or label names}

\item{samp_n}{Integer giving the number of rows in the output}

\item{numerical}{List of (quoted) variable names for which to append a numerical output, which
will be the underlying log odds for the discrete version of that variable.}
}
\description{
Generates samples from an ACY process. ACY refers to the paradigm where there
are multiple dicotomous variables and a causal DAG. In \code{acy_sim}, the DAG is implemented
by formulas, with the response variable on the left and the value to be calculated.
on the right.
}
\details{
The calculated values are log odds. In the output, these will be converted to
0,1 dicotomous variables. Exogenous variables should be set to a constant log odds.
(0 will give an equal prob of a 1/0 outcome.) Endogenous variables should be a mathematical
function of the exogenous and (sometimes) other endogenous variables.

A label name can be an element of \code{...} or \code{story}. This label will be used to denote
the observed value if the underlying variable is 1. (\code{o} or 0 is used otherwise.)
}
\examples{
acy_sim(x ~ 0, y ~ x, samp_n = 20, numerical = "y")
acy_sim(X ~ 0, Y ~ 10* X - 1, Z ~ - 3 + 10 * Y, X = 1, Y = "murder", Z = "telegraph", samp_n = 50)

}
