% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_thumbnail.R
\name{app_thumbnail}
\alias{app_thumbnail}
\title{Generate HTML (or Latex?) to show an image as a link to an app (or other webpage)}
\usage{
app_thumbnail(URL, img, caption = NULL)
}
\arguments{
\item{URL}{character string giving link to the app}

\item{img}{character string locating the image}

\item{caption}{character string setting the caption to display under the thumbnail}
}
\description{
Generate HTML (or Latex?) to show an image as a link to an app (or other webpage)
}
