% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{dag_system}
\alias{dag_system}
\title{Specify and simulate data from structural equation model}
\usage{
dag_system(...)
}
\arguments{
\item{...}{Formulas specifying the model.}
}
\description{
Create and run structural equation model using formulas for each node.
}
\details{
There cannot be any cycles in the DAG (as the name DAG indicates.) You can use \code{unif()}
and \code{gauss()} as random inputs.
}
\examples{
dag <- dag_system(A ~ unif(), B ~ 3 + 10*A)
dag_simulate(dag, n = 5)
}
