% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devoted_runners.R
\docType{data}
\name{Devoted_runners}
\alias{Devoted_runners}
\title{Race results for long-time runners of the Cherry Blossom 10-milers}
\format{A data.frame object with one row for each runner-year in the Cherry
Blossom 10-miler held in Washington, DC. Only those runners with 10 or more successive
participations in the race are included.
\itemize{
\item \code{age} the runner's age when the race  was run
\item \code{seconds} the time it took the runner to complete the 10-mile race in h:mm:ss
\item \code{sex}
\item \code{division} the sex/age division that the runner was in .
\item \code{year}
\item \code{Hometown}
\item \code{name}: the runners' name
\item \code{PiS/TiS} the runner's position among all runners of the same sex
\item \code{PiD/TiD}: the runner's position in his or her division.
}}
\usage{
data(Devoted_runners)
}
\description{
Race results for long-time runners of the Cherry Blossom 10-milers
}
\keyword{datasets}
