% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Energy_bills.R
\docType{data}
\name{Energy_bills}
\alias{Energy_bills}
\title{Utility bills over 20 years for a home}
\format{
A data.frame object with one row for each monthly bill.
\itemize{
\item \code{month}, \code{day}, \code{year} -- numerical values indicating the date the bill was produced
\item \code{temp} -- the average temperature for the period covered by the  bill, in degrees F
\item \code{kwh} -- the number of kilowatt-hours of electricity used
\item \code{ccf} -- the  amount of natural gas consumed, in cubic feet
\item \code{thermsPerDay} -- the average amount  of natural gas consumed each day, in a unit called "therms"
\item \code{dur} -- the number of days in  the period covered by the bill. Typically one month.
\item \code{totalbill} -- the total amount, in USD, of the bill
\item \code{gasbill} -- the part of \code{totalbill} attributable to natural gas use
\item \code{elecbill} -- the part of \code{totalbill} attributable to electricity consumption
}
}
\source{
The data were transcribed from  monthly utility bills for Danny Kaplan's house
in Saint Paul, Minnesota, USA
}
\usage{
data(Energy_bills)
}
\description{
Transcription of the energy bills received over a 20-year period from
1999 to 2019 for a residence in Saint Paul,  Minnesota, USA. The data are mostly
right, but there are some transcription errors.
}
\keyword{datasets}
