% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expiration.R
\docType{data}
\name{Expiration}
\alias{Expiration}
\title{Expiratory volume and smoking amount teens and pre-teens}
\format{A data.frame object with one row for each person.
\itemize{
\item \code{age} in years
\item \code{height} in  inches
\item \code{sex} ("male" or "female")
\item \code{volume} the full  expiratory lung volume (liters)
\item \code{smoke} whether the child or at least one parent  smokes.
}}
\source{
The data are  published as part of an article:
Michael Kahn,  "An Exhalent Problem for Teaching Statistics" (2005)
\emph{Journal of Statistical Education} Volume 13(2) \url{https://www.tandfonline.com/doi/full/10.1080/10691898.2005.11910559}
}
\usage{
data(Expiration)
}
\description{
These data are from studies aimed at assessing children's
pulmonary function in the absence or presence of smoking cigarettes,
as well as exposure to passive smoke from at least one parent.
They were published as part of  an article entitled "An
Exhalent Problem for Teaching Statistics" by Michael Kahn. (See source below.)
}
\keyword{datasets}
