% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_outlier.R
\name{gf_outlier}
\alias{gf_outlier}
\title{Plot the outliers of a variable}
\usage{
gf_outlier(...)
}
\arguments{
\item{...}{Like other gg_formula functions, the first argument can optionally be a gg object (usually via a pipe). There must always
be a formula, again as in other gg_formula commands.}

\item{size}{as in other ggplot2 geoms}

\item{color}{ditto}

\item{alpha}{ditto

This is just like gf_boxplot(), but doesn't draw the box!}
}
\description{
Plot the outliers of a variable
}
