% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homocides.R
\docType{data}
\name{Homocides}
\alias{Homocides}
\title{Country-by-country homocide data}
\format{A data.frame object with one row for each of about 287 countries in various years
from 1800 to 2015. (Not all countries are represented for all years. Similarly, some variables
have non-missing data for only a few years.) The unit of observation is a country in a year.
\itemize{
\item \code{country}: The name of the country
\item \code{year} the year of the observation
\item \code{population} the estimated population
\item \code{total_deaths} the total number of homicides. (Only for years 2002 and 2004.)
\item \code{murders_per_100k} the age-adjusted murder rate per 100,000 people
\item \code{rate_females} the female homocide rate, age adjusted, per 100,000 people
\item \code{rate_males} like \code{rate_females}
\item \code{rate_over_60} like \code{rate_females} but gives the rate for people over age 60 (males and females)
\item \code{rate_15_29} like \code{rate_over_60} but for people aged 15 to 29 years
\item \code{surface_area} the country's surface area in square kilometers
\item \code{birth_rate} the birth rate: the number of births per 1000 people
\item \code{population_15_19} count of the population aged 15 to 19 years
\item \code{female_15_19} the fraction of the population (\%) that is female and aged 15 to 19
\item \code{male_15_19} the fraction of the population (\%) that is male and aged 15 to 19
\item \code{male_unemployment} the fraction of the male workforce that is unemployed
\item \code{youth_unemployment} the fraction of youth (how defined?) that is not in school or employed
}}
\source{
The data themselves were scraped from the \href{https://www.gapminder.org/data/}{Gapminder site} on July 25, 2019. Gapminder
organizes the data in separate columns for each year. These have been gathered together so that
the unit of observation is a country in a year.
}
\usage{
data(Homocides)
}
\description{
Country-by-country homocide data
}
\details{
The data have been put together from \url{https://www.gapminder.org/data/}
}
\keyword{datasets}
