% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_thumbnail_image.R
\name{make_thumbnail_image}
\alias{make_thumbnail_image}
\title{Create a thumbnail image of a URL}
\usage{
make_thumbnail_image(URL, delay = 25)
}
\arguments{
\item{URL}{character string with the URL of the app}

\item{delay}{number of seconds to delay before snapping the thumbnail}
}
\description{
Create a thumbnail image of a URL
}
