% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS_table.R
\name{margin_note}
\alias{margin_note}
\title{A margin note that will work in gitbook or tufte}
\usage{
margin_note(text, icon = "[Click to see note.]")
}
\description{
A margin note that will work in gitbook or tufte
}
