% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montana_poll.R
\docType{data}
\name{Montana_poll}
\alias{Montana_poll}
\title{Data from an online poll}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 573 rows and 10 columns.}
\usage{
Montana_poll
}
\description{
It's hard to get pollsters' raw data; only the tabulated results
seem to be available. This is some [actual data](https://surveys.google.com/reporting/survey?survey=mqbskxyva7gb4h4uwgc7j6h2na),
which relate to a 2017 Montana special election for its Congressional seat.
We found the data by following the link, which was published
[in this story at fivethirtyeight.com](https://fivethirtyeight.com/features/fake-polls-are-a-real-problem/).
}
\keyword{datasets}
