% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QB_draft.R
\docType{data}
\name{QB_draft}
\alias{QB_draft}
\title{Quarterbacks in the National Football League draft, 1997-2013}
\format{A data.frame object with one row for each American football college quarterback
who was drafted into the National Football League.
\itemize{
\item \code{Year}: Year drafted into the NFL
\item \code{Round}: Draft round the player was selected
\item \code{Pick}: Order of the player among those drafted in his year
\item \code{NFLTeam}: The NFL team that drafted the player
\item \code{College}: The college played for
\item \code{Div1A}: Whether the college was in Division 1A of the NCAA
\item \code{Name}: Player's name
\item \code{ColG}: Number of college games played
\item \code{ColRushAtt}: Number of rush attempts made in college games
\item \code{ColRushYds}: Total yards gained in those rush attempts
\item \code{ColRushTD}: Number of touchdowns scored in those rush attempts
\item \code{ColAtt}: Number of passes attempted in college games
\item \code{ColComp}: Number of completed passes in college games
\item \code{ColInt}: Number of intercepted passes given up in college games
\item \code{ColYds}: Total yards gained in college games
\item \code{ColTD}: Total number of
\item \code{G} - Number of NFL games played
\item \code{Comp}: Number of passes completed in NFL games
\item \code{Att}: Number of passes attempted in NFL games
\item \code{Yds}: Number of yards gained in NFL games
\item \code{TD}: Number of touchdowns scored in NFL games
\item \code{Int}: Number of intercepted passes given up in NFL games
\item \code{RushAtt}: Number of rushing attempts in NFL games
\item \code{RushYds}: Total yards gained in those attempts
\item \code{RushTD}: Number of times touchdown scored on a
\item \code{CarAV}: "Career approximate value" statistic for player
\item \code{Sacks}: Number of times sacked in NFL games
\item \code{FumLost}: Number of fumbles made in NFL games
\item \code{ProBowls}: Number of NFL pro-bowl games played
\item \code{YrsStarterPrior2015}: Numbe of years the QB was a starter for their pro team (before 2015, when the data were assembled).
\item \code{Wonderlic}: Score on Wonderlic intelligence test
\item \code{height}: Height in inches
\item \code{weight}: Weight in pounds
\item \code{BenchReps}: Number of bench presses
\item \code{Yard40}: Time to run 40 yards (secs)
\item \code{Yard20}: Time to run 20 yards (secs)
\item \code{Yard10}: Time to run 10 yards (secs)
\item \code{VerticalJump}: Height of vertical jump (inches)
\item \code{BroadJump}: Length of broad jump (inches)
\item \code{shuttle}: time to complete the shuttle shuttle (secs)
\item \code{cone}: time to complete the 3-cone drill (secs)
\item \code{time}: the record time in seconds
\item \code{swimmer} the name of the swimmer
\item \code{date} a Date object containing the date the record was made
\item \code{place} string descripting the location
\item \code{sex}: coded as \code{F} and \code{M}
\item \code{lengths}: the total distance was divided into lengths of either 25 or 50 meters. \code{lengths} gives the number of such lengths in the total distance.
\item \code{dist}: the total distance (in meters) of the race
}}
\source{
Wolfson, J., Addona, V., and Schmicker, R. (2017) Forecasting the performance
of college prospects selected in the National Football League draft. In J.
Albert, M.E. Glickman, T.B. Swartz, & R.H. Koning (Eds.), Handbook of Statistical
Methods and Analyses in Sports. Chapman & Hall/CRC.
}
\usage{
data(QB_draft)
}
\description{
Quarterbacks in the National Football League draft, 1997-2013
}
\keyword{datasets}
