% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanks.R
\name{question_blank}
\alias{question_blank}
\title{Blanks and spaces for student answers in documents}
\usage{
question_blank(prompts = "", length = 3, underline = "  .", breaks = 2)
}
\arguments{
\item{prompts}{vector of character strings to prompt the answer}

\item{length}{length of blank in mm}

\item{underline}{the short character string to use in composing the underline.}

\item{breaks}{number of line breaks following each prompt
Default: nonbreaking space followed by a period.}
}
\description{
Blanks and spaces for student answers in documents
}
