% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Race_weather.R
\docType{data}
\name{Race_weather}
\alias{Race_weather}
\title{Weather during the Cherry Blossom 10-miler}
\format{
A data.frame object with one row for each year the  Cherry Blossom 10-miler
was held
\itemize{
\item \code{year}
\item \code{month} Race is held on a Sunday in early April
\item \code{day}
\item \code{temp} Temperature in deg. F
\item \code{humidity} Humidity in percent
\item \code{wind_speed} Average wind speed
\item \code{gust} Speed of wind gusts, if substantially different from \code{wind_speed}
\item \code{precip} Precipitation during one hour, in  inches
\item \code{clouds} Description of the sky
}
}
\source{
Scraped from Weather Underground via URLs like
\url{https://www.wunderground.com/history/daily/us/dc/washington/KDCA/date/2010-4-11}
}
\usage{
data(Race_weather)
}
\description{
Weather during the Cherry Blossom 10-miler
}
\keyword{datasets}
