% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramsey_weather.R
\docType{data}
\name{Ramsey_weather}
\alias{Ramsey_weather}
\title{Hourly weather records in Ramsey County, MN}
\format{A data.frame object with one row for measurement time
\itemize{
\item \code{usaf_station}: the USAF code for the weather station (constant)
\item \code{wban_station}: the WBAN code for the weather station (constant)
\item \code{date_time}: a POSIXct format date & time
\item \code{latitude}: constant
\item \code{longitude}: constant
\item \code{wind_speed}: in miles-per-hour
\item \code{wind_direction}: direction wind is coming from, as a compass reading: 0 - 360 degrees, with 0 meaning north
\item \verb{air pressure}: in hectopascals
\item \code{temperature_dewpoint}: in degrees C
\item \code{temperature}: in degrees C
\item \code{precip}: amount of liquid precipitation, in mm, over the last interval
\item \code{precip_hrs}: the length of the interval over which \code{precip} was accumulated
\item \code{precip_rate}: a rate: \code{precip / precip_hrs}.
}}
\source{
See the script in \code{system.file(package = "SDSdata", "raw_data/weather/read_weather.R")}
}
\usage{
data(Ramsey_weather)
}
\description{
These data, covering the period from late September 1992 through mid July 2018, were scraped from
the NOAA weather site. Measurements are made roughly every hour, although there are gaps and occasions
when the measurements are much more frequent.
}
\details{
Some of the precipitation measurements overlap in time. For instance, there might be a 6-hour interval over which
precipitation was summed, as well as 1-hour intervals containing precipation counted in the 6-hour interval. Use \code{precip_rate} as
an indicator of how hard it was raining during the interval covered.
}
\keyword{datasets}
