% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schrek_1950.R
\docType{data}
\name{Schrek_1950}
\alias{Schrek_1950}
\title{Smoking among cancer patients}
\format{
\code{Schrek_1950}: A data.frame object with one row for each patient.
\itemize{
\item \code{age}: The patient's age. This  is rounded down to the nearest decade, since Schrek et al. presented
ages  as 20, 30, 40, 50, ...
\item \code{site}: The site of the patient's tumor: "lung", "lip", "larynx", "other
\item \code{smoker}: Whether the patient smoked: "smoker", "nonsmoker". More accurate names for the
levels would be "moderate or heavy smoker" and "nonsmoker or  light smoker." Light smoking
was defined as consuming 10 or fewer cigarettes per day (or, two or fewer cigars or three  or fewer pipe
bowls).
}
}
\source{
Reconstructed from Schrek, Baker, et al.,
"Tobacco Smoking as an Etiologic Factor in Disease. I. Cancer"
Cancer Research 1950;10:49-58.
<cancerres.aacrjournals.org/content/10/1/49>
}
\usage{
data(Schrek_1950)
}
\description{
Smoking among cancer patients
}
\details{
\code{Schrek_1950} is a reconstruction of the "raw" data behind the
summary presented in  Schrek et al. (1950) "Tobacco Smoking as an Etiologic Factor in Disease. I. Cancer"
Cancer Res,10:49-58. That case-control study compared smoking rates among white men with tumors
at different sites. Since the interest was in testing the  hypothesis that lung, lip, and/or larynx
cancer is associated with smoking, the control group was those patients with "other" tumors. This choice
of control group was motivated by the availability of subjects and the desire to make the
control group as similar as possible to the case group.
}
\keyword{datasets}
