% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_sim.R
\name{scm_sim}
\alias{scm_sim}
\title{Simulate structural equation processes}
\usage{
scm_sim(story = NULL, ..., samp_n = 5)
}
\arguments{
\item{story}{Optional list containing formulas and/or label names}

\item{...}{more formulas describing the process and/or label names}

\item{samp_n}{Integer giving the number of rows in the output}
}
\description{
Generates samples from a structural causal model (SCM).  In \code{scm_sim}, the DAG is implemented
by formulas, with the response variable on the left and the value to be calculated.
on the right.
}
\details{
You can set the exogenous random component of any variable by using
the desired \code{r____()} random number generator with \code{samp_n} as the value of size.

A label name can be an element of \code{...} or \code{story}. This label will be used to denote
the observed value if the underlying variable is 1. (\code{o} or 0 is used otherwise.)
}
\examples{
scm_sim(X ~ rnorm(samp_n), Y ~ 10 + runif(samp_n) + 5 * X, samp_n = 20)
scm_sim(X ~ 0, Y ~ 10* X - 1, Z ~ - 3 + 10 * Y, X = 1, Y = "murder", Z = "telegraph", samp_n = 50)

}
