% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_character.R
\name{sds_character}
\alias{sds_character}
\title{Render a special character in either html or latex as appropriate.}
\usage{
sds_character(char, show = FALSE,
  format = ifelse(knitr::is_latex_output(), "latex", "html"))
}
\arguments{
\item{char}{the character, as a quoted string}

\item{show}{if TRUE, return a vector showing the available characters. Not intended for use in documents, just to help identify
when a character you want to use is missing.}

\item{format}{either \code{"html"} or \code{"latex"}. Default uses \code{knitr::is_latex_output()}}
}
\description{
Given a character from a special unicode or other font, translate it into a tex command if needed.
}
\details{
The point of this function is to let you use native characters in html mode, but avoid characters that
latex doesn't recognize by using the corresponding latex command when in latex mode. Why not just use the latex command
in both cases. That's not a bad practice, but it causes problems when trying to create a Kindle edition, since Kindle doesn't
process latex.
}
\examples{
sds_character("♣", format = "latex")
sds_character(show = TRUE)

}
