% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_image.R
\name{sds_image}
\alias{sds_image}
\title{Render an image in either html or latex as appropriate.}
\usage{
sds_image(img, width = NULL, height = NULL, lwidth = width,
  lheight = height, format = ifelse(knitr::is_latex_output(), "latex",
  "html"))
}
\arguments{
\item{img}{quoted relative path name of image}

\item{width}{number giving width as percent}

\item{height}{number giving height as percent}

\item{lwidth}{number for latex width, if you want a different width for latex and for html}

\item{lheight}{number for latex height, like \code{lwidth}}

\item{format}{either \code{"html"} or \code{"latex"}. Default uses \code{knitr::is_latex_output()}}
}
\description{
Insert an image as either html or latex
}
\examples{
sds_image("images/bogus", format = "latex", width = 50)

}
