% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_latex.R
\name{sds_latex}
\alias{sds_latex}
\title{Utilities for combining  latex environments into markdown}
\usage{
sds_latex(environment, position = c("begin", "end"))
}
\arguments{
\item{environment}{The name of the latex environment, e.g. \code{"fullwidth"}}

\item{position}{Either \verb{"begin"`` or  }"end"` to mark the beginning or end of an environment}
}
\description{
Utilities for combining  latex environments into markdown
}
\examples{
sds_latex("fullwidth", "begin")

}
