% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_newword.R
\name{sds_newword}
\alias{sds_newword}
\title{Define a new word in the text.}
\usage{
sds_newword(str, format = ifelse(knitr::is_latex_output(), "latex",
  "html"))
}
\arguments{
\item{str}{the character string for the new word}

\item{format}{either \code{"html"} or \code{"latex"}. Default uses \code{knitr::is_latex_output()}}
}
\description{
Given a character from a special unicode or other font, translate it into a tex command if needed.
}
\examples{
sds_newword("log likelihood", format = "latex")

}
