% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_setup.R
\name{sds_setup}
\alias{sds_setup}
\title{Set up script for documents in the SDS book}
\usage{
sds_setup(...)
}
\arguments{
\item{...}{additional arguments. Not used currently.}
}
\description{
Loads various packages and sets parameters.
}
