% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDS_table.R
\name{SDS_table}
\alias{SDS_table}
\title{Format data frames nicely for SDS book}
\usage{
SDS_table(data, show_n = 6L, nrows = nrow(data), format = "html",
  caption = "", ...)
}
\arguments{
\item{data}{the data frame to be displayed}

\item{show_n}{how many lines to display (default 6. Use `Inf` for all rows.)}

\item{nrows}{an integer specifying how many rows the from the data frame should be displayed in the footnote. Set to `Inf` to suppress footnote.}

\item{format}{either `"html"` or `"latex"`}

\item{caption}{string to pass as a caption to the table}

\item{...}{additional arguments to `kableExtra::kable_styling()`, e.g. `"striped"`, `"hover"`}
}
\description{
Creates HTML or Latex markup with a footnote showing how many rows
there are in total.
}
\examples{
SDS_table(mtcars, "striped", "hover")
SDS_table(mtcars, show_n = 3, caption="Data about cars")
}
