% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_table_caption.R
\name{sds_table_caption}
\alias{sds_table_caption}
\title{Adds a table caption for a markdown table or other content}
\usage{
sds_table_caption(
  caption,
  name = NULL,
  format = ifelse(knitr::is_latex_output(), "latex", "html")
)
}
\arguments{
\item{caption}{the character string for the new word}

\item{name}{character string specifying reference label for the caption}

\item{format}{either \code{"html"} or \code{"latex"}. Default uses \code{knitr::is_latex_output()}}
}
\description{
Adds a table caption for a markdown table or other content
}
\examples{
sds_table_caption("This can be the table caption!")

}
