% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Titanic_crew.R
\docType{data}
\name{Titanic_crew}
\alias{Titanic_crew}
\title{Crew on the Titanic}
\format{\code{Titanic_crew}: A data.frame object with one row for each crew member on the ship when she sank:
\itemize{
\item \code{crew}: The broad category of job held by the crewman
\item \code{job}: A more specific description of the job
\item \code{sex}: All the crew were male
\item \code{survived}: Whether the crewmember survived the sinking. If so, 1. Otherwise, 0.
\item \code{age}: The crewman's age, in years
\item \code{boat}: For survivors, the lifeboat from which they were picked up
\item \code{body}: For those whose body was recovered, the number assigned to the body. Missing for survivors or those whose body was never recovered.
}}
\source{
Scraped from \url{https://www.encyclopedia-titanica.org/titanic-passenger-lists/}.
}
\usage{
data(Titanic_passengers)
}
\description{
Crew on the Titanic
}
\details{
\code{Titanic_crew} is a roster of 895 crew members and officers believed to be on the Titanic when she sank.
}
\keyword{datasets}
