% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Titanic_passengers.R
\docType{data}
\name{Titanic_passengers}
\alias{Titanic_passengers}
\title{Passengers and crew on RMS Titanic}
\format{\code{Titanic_passengers}: A data.frame object with one row for each passenger
\itemize{
\item \code{class}: The passenger's cabin class. Values first, second, or steerage.
\item \code{survived}: Whether the passenger survived the sinking. Values: 1 if survived, 0 otherwise
\item \code{sex}: the passenger's sex, \code{female} or \code{male}
\item \code{age}: the age of the passenger (in years). Infants under 1 have a fractional year, others are truncated or rounded
to the nearest integer.
\item \code{sibsp}: the number of siblings (including the spouse, if any) travelling with the passenger
\item \code{parch}: the number of parents or children travelling with the passenger.
\item \code{fare}: the fare paid for the passenger's ticket, GBP
\item \code{embarked}: where the passenger got on to the Titanic: Southhampton, Cherbourg, Queenstown
\item \code{boat}: the life boat into which the passenger was first placed.  A handful  of passengers
were transferred from one boat to another. This  variable is the first boat.
\item \code{body}: the number assigned when the body was recovered. Empty for survivors or those whose body was never found.
\item \code{cabin}: the passenger's cabin number
\item \code{ticket}: the passenger's ticket number. (Passengers travelling together might have shared a ticket.)
\item \code{destination}: the home or destination town for the passenger
\item \code{name}: The passenger's name
}}
\source{
\url{http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/titanic3.xls} There are a number of
other potential sources for passenger lists, for instance \url{https://www.encyclopedia-titanica.org/titanic-passenger-lists/}.
}
\usage{
data(Titanic_crew)
}
\description{
Passengers and crew on RMS Titanic
}
\details{
The \code{Titanic_passengers} data frame is a roster of 1309 passengers on the steamship Titanic.
At the time of the sinking, there were approximately 2400 people on board the ship. Of these, approximately
1330 were passengers, together with approximately 895 crew. The passenger list includes approximately
42 people who were servants of other passengers. According to many reports, 712 people
survived  the sinking. (See, e.g. \url{https://www.encyclopedia-titanica.org/titanic-survivors/}.)
The \code{survived} column,  combining \code{Titanic_passengers} and \code{Titanic_crew} lists only 628 survivors. The source
of the  discrepancy is unknown.
}
\seealso{
Titanic_crew
}
\keyword{datasets}
