% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vaccine_sim.R
\docType{data}
\name{Vaccine_sim}
\alias{Vaccine_sim}
\title{Simulated vaccine/flu/mortality data}
\format{An object of class \code{data.frame} with 10000 rows and 4 columns.}
\usage{
Vaccine_sim
}
\description{
These data were generated with a structural causal model. See details.
}
\details{
A data frame object with one row for each (simulated) person:
\itemize{
\item \code{health}: whether the person was in poor health
\item \code{vaccination} did the person get vaccinated
\item \code{flu} did the person get the flu
\item \code{mortality}: did the person die
}

The data were generated with
\code{set.seed(102)}

\code{acy_sim(Health ~ 0,}
\itemize{
\item \preformatted{    `Vaccination ~ 1.73 - 3.47 * Health,`
}
\item \preformatted{     `Flu ~ log(.21/.79) + (log(.2/.8) - log(.21/.79)) * Vaccination,`
}
\item \preformatted{     `Death ~ -4.595 + 0.704*Flu + 2.86*Health -  0.3*Health*Flu,`
}
\item \preformatted{     `Health = "Poor", Vaccination = "Received", Flu = "Yes",``
}
\item \preformatted{     `Mortality = "Died", samp_n = 10000)`
}
}
}
\keyword{datasets}
