% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{include_exercise}
\alias{include_exercise}
\title{Insert an exercise from a file}
\usage{
include_exercise(id, show_answer = TRUE, directory = "Exercises/")
}
\description{
Each exercise is contained in its own file, with a unique ID. This function
reads the file for that ID, knits the contents, and returns the knitted contents.
Typically this will be used to insert the exercise into an Rmd file.
}
