% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds_table.R
\name{sds_table}
\alias{sds_table}
\title{Format data frames nicely for SDS book}
\usage{
sds_table(data, show_n = 6L, nrows = nrow(data), footnote = NULL,
  format = ifelse(knitr::is_latex_output(), "latex", "html"),
  caption = "", align = NULL, ...)
}
\arguments{
\item{data}{the data frame to be displayed}

\item{show_n}{how many lines to display (default 6. Use \code{Inf} for all rows.)}

\item{nrows}{an integer specifying how many rows the from the data frame should be displayed in the footnote. Set to \code{Inf} to suppress footnote.}

\item{footnote}{character string to be placed at the bottom of the table in place of the usual "... and so on for __ rows."}

\item{format}{either \code{"html"} or \code{"latex"}. Default will use \code{knitr::is_latex_output()} to decide.}

\item{caption}{string to pass as a caption to the table}

\item{align}{optional column alignment instructions passed to \code{kable}}

\item{...}{additional arguments to \code{kableExtra::kable_styling()}, e.g. \code{"striped"}, \code{"hover"}}
}
\description{
Creates HTML or Latex markup with a footnote showing how many rows
there are in total.
}
\examples{
sds_table(mtcars, "striped", "hover")
sds_table(mtcars, show_n = 3, caption="Data about cars")
}
